Move and enhance the server hook to inject metadata, seems promising, getting first good results in days
